backend/transactions/requirements.txt
boto3==1.26.93
Flask==2.0.2
Flask-Cors==3.0.10
python-dotenv==0.19.2